Add/amend specs and feature tests
Change tests to reflect the change in flow to include the new 'About the
financial assessment screen'.
